1. Move VAD-related functions into vad.c/vad.h 2. Rename Unicode string function parse_unicode_str -> parse_unicode_strptr    parse_unicode_strptr() parse the string by the string address 3. Rename Unicode string function parse_unicode_str_new -> parse_unicode_str    parse_unicode_str() parse the string by the string content
