+++ ext/memfrs/memfrs.h
@@ -28,6 +28,7 @@
 #include "include/utarray.h"
 
 #define STRLEN 128
+#define SIZEOFUNICODESTRING 0x10
 
 //represent the field in the data structure
 typedef struct field_info
@@ -41,12 +42,13 @@ typedef struct field_info
 
 extern uint64_t g_kpcr_ptr;
 
+extern bool memfrs_check_struct_info(void);
 extern int memfrs_load_structs( const char* type_filename);
 extern bool memfrs_kpcr_self_check( uint64_t seg_gs_cpl0 );
 extern int  memfrs_enum_proc_list( uint64_t seg_gs_cpl0, CPUState *cp );
-extern void traverse_vad_tree(uint64_t eprocess_adr, CPUState *cpu);
-extern void parse_unicode_str(uint64_t ustr_ptr, CPUState *cpu);
-extern void parse_unicode_str_new(uint8_t* ustr, CPUState *cpu);
+//extern void traverse_vad_tree(uint64_t eprocess_adr, CPUState *cpu);
+extern void parse_unicode_strptr(uint64_t ustr_ptr, CPUState *cpu);
+extern void parse_unicode_str(uint8_t* ustr, CPUState *cpu);
 extern json_object* memfrs_q_struct(const char* ds_name);
 extern field_info* memfrs_q_field( json_object* struc, const char* field_name  );
 extern int memfrs_close_field(field_info* field);
